Expose pure-logic helpers for direct unit testing
Make currently-internal helpers importable so each team member can own
unit tests for them without going through their public callers. Paired
with docs/unit-test-assignments.md, which splits the work across the
five team members.

 - Export mergeParameters, buildParameters, buildRequestBody,
   buildResponses, and extractSwagger2Body from parseEndpoints.js
 - Export populateSchemaUsage from parseSpec.js
 - Export setHeader from buildRequest.js
 - Export buildInitialRequestBody from useRequestPipeline.js
 - Move statusCodeClass, prettyPrintXml, formatResponseBody, and the
   XML escape helpers into a new formatResponse.js so they can be
   tested without rendering ResponseDisplay; retarget the existing
   responseDisplay.test.js import at the new module
 - Add @testing-library/react and @testing-library/dom as dev deps
   for renderHook/act when testing useRequestPipeline
 - Drop unused tags destructure in MetadataView and unused minimalSpec
   helper in validateSpec.test.js to clear the lint baseline
